backend/requirements.txt
fastapi
uvicorn[standard]
mangum
python-dotenv
transformers
spacy
bert-extractive-summarizer